+++ src/plugins/XKTLoaderPlugin/XKTLoaderPlugin.js
@@ -382,10 +382,67 @@ parsers[ParserV6.version] = ParserV6;
  * myViewer.scene.setObjectVisibilities("myModel1#0BTBFw6f90Nfh9rP1dlXrb", true);
  *````
  *
- * ## BCFViewpointsPlugin Consequences
+ * ## Validating the Metadata
  *
- * Using ````XKTLoaderPlugin```` to load models with {@link XKTLoaderPlugin#globalizeObjectIds} ````true```` has consequences
- * for BCF viewpoints created by {@link BCFViewpointsPlugin#getViewpoint}.
+ * As mentioned, XKTLoaderPlugin can also load an accompanying JSON metadata file with each model, which creates a
+ * {@link MetaModel} corresponding to the model {@link Entity} and a {@link MetaObject} corresponding to each
+ * object {@link Entity}.
+ *
+ * When loading metadata, we have the option to supply a callback to validate it. Ideally, our metadata should
+ * be fully validated further up the data pipeline, but this callback gives us one more chance to ensure that the
+ * metadata is going to work properly with other functionality, such as that provided by {@link TreeViewPlugin}.
+ *
+ * The example below shows how we can use the callback to quickly check if our metadata is going to work properly
+ * with {@link TreeViewPlugin}, and abort the load:
+ *
+ * ````javascript
+ * import {
+ *   isModelValidForContainmentHierarchy,
+ *   isModelValidForStoreysHierarchy,
+ *   isModelValidForTypesHierarchy
+ * } from "../src/plugins/TreeViewPlugin/modelValidation.js";
+ *
+ * const metaModelValidationErrors = [];
+ *
+ * const model = xktLoader.load({
+ *
+ *      id: "Widget",
+ *      src: "./models/xkt/schependomlaan/schependomlaan.xkt",
+ *      metaModelSrc: "./metaModels/schependomlaanWithErrors/metaModelWithoutIfcBuilding.json",
+ *      edges: true,
+ *
+ *      validateMetaModel: (metaModel) => {
+ *          const valid =
+ *              validateMetaModelForTreeViewContainmentHierarchy(metaModel, metaModelValidationErrors) &&
+ *              validateMetaModelForTreeViewStoreysHierarchy(metaModel, metaModelValidationErrors) &&
+ *              validateMetaModelForTreeViewTypesHierarchy(metaModel, metaModelValidationErrors);
+ *          return valid;
+ *      }
+ *  });
+ *
+ * model.on("loaded", () => {
+ *      const t1 = performance.now();
+ *      document.getElementById("time").innerHTML = "Model loaded in " + Math.floor(t1 - t0) / 1000.0 + " seconds<br>Objects: " + model.numEntities;
+ *  });
+ *
+ * model.on("error", (e) => {
+ *
+ *      if (metaModelValidationErrors.length > 0) {
+ *
+ *          // We only get validation errors if all else went OK and there
+ *          // are no other errors (ie. the metamodel JSON loaded and parsed OK)
+ *
+ *          console.error("Error loading model - metadata validation errors:<br><br> " + metaModelValidationErrors.join("<br>"));
+ *
+ *      } else {
+ *          console.error("Error loading model: " + e);
+ *      }
+ *
+ *      model.destroy();
+ * });
+ * ````
+ *
+ * * [[Run example](https://xeokit.github.io/xeokit-sdk/examples/#loading_XKT_validateMetaModel)]
  *
  * @class XKTLoaderPlugin
  */
@@ -401,11 +458,18 @@ class XKTLoaderPlugin extends Plugin {
      * @param {Object} [cfg.dataSource] A custom data source through which the XKTLoaderPlugin can load model and metadata files. Defaults to an instance of {@link XKTDefaultDataSource}, which loads uover HTTP.
      * @param {String[]} [cfg.includeTypes] When loading metadata, only loads objects that have {@link MetaObject}s with {@link MetaObject#type} values in this list.
      * @param {String[]} [cfg.excludeTypes] When loading metadata, never loads objects that have {@link MetaObject}s with {@link MetaObject#type} values in this list.
-     * @param {Boolean} [cfg.excludeUnclassifiedObjects=false] When loading metadata and this is ````true````, will only load {@link Entity}s that have {@link MetaObject}s (that are not excluded). This is useful when we don't want Entitys in the Scene that are not represented within IFC navigation components, such as {@link StructureTreeViewPlugin}.
+     * @param {Boolean} [cfg.excludeUnclassifiedObjects=false] When loading metadata and this is ````true````, will only load {@link Entity}s that have {@link MetaObject}s (that are not excluded). This is useful when we don't want Entitys in the Scene that are not represented within IFC navigation components, such as {@link TreeViewPlugin}.
+     * @param {Number} [cfg.maxGeometryBatchSize=50000000] Maximum geometry batch size, as number of vertices. This is optionally supplied
+     * to limit the size of the batched geometry arrays that {@link PerformanceModel} internally creates for batched geometries.
+     * A low value means less heap allocation/de-allocation while loading batched geometries, but more draw calls and
+     * slower rendering speed. A high value means larger heap allocation/de-allocation while loading, but less draw calls
+     * and faster rendering speed. It's recommended to keep this somewhere roughly between ````50000```` and ````50000000```.
      */
     constructor(viewer, cfg = {}) {
 
         super("XKTLoader", viewer, cfg);
+
+        this._maxGeometryBatchSize = cfg.maxGeometryBatchSize;
 
         this.dataSource = cfg.dataSource;
         this.objectDefaults = cfg.objectDefaults;
@@ -605,6 +669,7 @@ class XKTLoaderPlugin extends Plugin {
      * @param {Boolean} [params.backfaces=false] Indicates if backfaces are visible on the model. Making this ````true```` will reduce rendering performance.
      * @param {Boolean} [params.excludeUnclassifiedObjects=false] When loading metadata and this is ````true````, will only load {@link Entity}s that have {@link MetaObject}s (that are not excluded). This is useful when we don't want Entitys in the Scene that are not represented within IFC navigation components, such as {@link TreeViewPlugin}.
      * @param {Boolean} [params.globalizeObjectIds=false] Indicates whether to globalize each {@link Entity#id} and {@link MetaObject#id}, in case you need to prevent ID clashes with other models. See {@link XKTLoaderPlugin#globalizeObjectIds} for more info.
+     * @param {Function} [params.validateMetaModel] Optional callback to validate the {@link MetaModel}, before loading the *````.xkt````*. The callback should return ````true```` if the MetaModel is valid, otherwise ````false````, in which case this function then aborts loading and returns an empty {@link PerformanceModel}, on which it fires an "error" event.
      * @returns {Entity} Entity representing the model, which will have {@link Entity#isModel} set ````true```` and will be registered by {@link Entity#id} in {@link Scene#models}.
      */
     load(params = {}) {
@@ -615,7 +680,8 @@ class XKTLoaderPlugin extends Plugin {
         }
 
         const performanceModel = new PerformanceModel(this.viewer.scene, utils.apply(params, {
-            isModel: true
+            isModel: true,
+            maxGeometryBatchSize: this._maxGeometryBatchSize
         }));
 
         const modelId = performanceModel.id;  // In case ID was auto-generated
@@ -656,13 +722,18 @@ class XKTLoaderPlugin extends Plugin {
 
             const processMetaModelData = (metaModelData) => {
 
-                this.viewer.metaScene.createMetaModel(modelId, metaModelData, {
+                const metaModel = this.viewer.metaScene.createMetaModel(modelId, metaModelData, {
                     includeTypes: includeTypes,
                     excludeTypes: excludeTypes,
-                    globalizeObjectIds: this.globalizeObjectIds
+                    globalizeObjectIds: this.globalizeObjectIds,
+                    validateMetaModel: params.validateMetaModel
                 });
 
                 this.viewer.scene.canvas.spinner.processes--;
+
+                if (!metaModel) {
+                    return false;
+                }
 
                 if (params.src) {
                     this._loadModel(params.src, params, options, performanceModel);
@@ -673,6 +744,8 @@ class XKTLoaderPlugin extends Plugin {
                 performanceModel.once("destroyed", () => {
                     this.viewer.metaScene.destroyMetaModel(performanceModel.id);
                 });
+
+                return true;
             };
 
             if (params.metaModelSrc) {
@@ -683,19 +756,36 @@ class XKTLoaderPlugin extends Plugin {
 
                 this._dataSource.getMetaModel(metaModelSrc, (metaModelData) => {
 
-                    this.viewer.scene.canvas.spinner.processes--;
+                    if (performanceModel.destroyed) {
+                        return;
+                    }
 
-                    processMetaModelData(metaModelData);
+                    if (!processMetaModelData(metaModelData)) {
+
+                        this.error(`load(): Failed to load model metadata for model '${modelId} from '${metaModelSrc}' - metadata not valid`);
+
+                        performanceModel.fire("error", "Metadata not valid");
+                    }
+
+                    this.viewer.scene.canvas.spinner.processes--;
 
                 }, (errMsg) => {
 
                     this.error(`load(): Failed to load model metadata for model '${modelId} from  '${metaModelSrc}' - ${errMsg}`);
 
+                    performanceModel.fire("error", `Failed to load model metadata from  '${metaModelSrc}' - ${errMsg}`);
+
                     this.viewer.scene.canvas.spinner.processes--;
                 });
 
             } else if (params.metaModelData) {
-                processMetaModelData(params.metaModelData);
+
+                if (!processMetaModelData(params.metaModelData)) {
+
+                    this.error(`load(): Failed to load model metadata for model '${modelId} from '${metaModelSrc}' - metadata not valid`);
+
+                    performanceModel.fire("error", "Metadata not valid");
+                }
             }
 
         } else {
@@ -728,6 +818,10 @@ class XKTLoaderPlugin extends Plugin {
 
     _parseModel(arrayBuffer, params, options, performanceModel) {
 
+        if (performanceModel.destroyed) {
+            return;
+        }
+
         const dataView = new DataView(arrayBuffer);
         const dataArray = new Uint8Array(arrayBuffer);
         const xktVersion = dataView.getUint32(0, true);
@@ -754,6 +848,9 @@ class XKTLoaderPlugin extends Plugin {
         performanceModel.finalize();
 
         performanceModel.scene.once("tick", () => {
+            if (performanceModel.destroyed) {
+                return;
+            }
             performanceModel.scene.fire("modelLoaded", performanceModel.id); // FIXME: Assumes listeners know order of these two events
             performanceModel.fire("loaded", true, false); // Don't forget the event, for late subscribers
         });